requirements.txt
aiogram==3.15.0
fastapi==0.115.0
uvicorn[standard]==0.30.0
asyncpg==0.30.0
httpx==0.27.0
python-dotenv==1.0.1
